refractor: renaming the instance segmentation module for proper imports
